Core: Event and event Engine.
The Event class is an abstract class, intended to be subclassed by
specific use cases. It is a simple class that holds some timing
information and context and yields other Events when called. Subclasses
just have to implment a `call` method that yields either None or a
series of Events to be scheduled.

The Engine has the primary function to schedule and consume events from
a priority queue.  The queue expects all items to be a subclass of Event
and is prioritized by the event timestamp.  Additionally, it manages an
iternal status, which can be controlled by events raising specific
custom exceptions.

If an Event raises a StopEngineError, the engine will stop gracefully in
place.  If an Event raises any other subclass of EventError, it indicates a core
issue and signals the engine to abort.

This is the first real bit of code to the project and I am happy with
some of the new development choices and architectures being explored.
Onwards and upwards!