Small touchups to Event and Engine.
The typing setup on Event was wrong.  It was more correct to simply use a forward referencing type for the event callbacks.

The Engine class has some refactor of the logic in the run method.  There is a new method to consume the event and the engine timestamp was updated at a more correct position.