Chart of Accounts added.
This is an account management object that holds relevant accounts for
the simulation.  It organizes the accounts in a dict with key structure:

`accounts[ACCOUNT_TYPE][ACCOUNT_NAME]`

where both `ACCOUNT_TYPE` and `ACCOUNT_NAME` are strings.

There are convenience methods for adding, removing, and checking
existence of an account.  Most methods allow for passing account and
 account_type by a string of the name or a direct object.

 There are also some filtering methods to help retrieve relevant data
 retrievals.