Refactored Event to Engine module
Also removed some dataclass assignments on some classes, it was
interfering with some inheritance structures and is not really necessary
outside of pretty repr methods

The Engine also manages QueueItem in the queue which hold a timestep and
an event.  The class is ordered by the timestep.

These changes were all in response to the first real integration test.
Likely more to follow